Add dynamic batch size

+++ scripts/finetune_superglue_grid.sh
@@ -4,41 +4,41 @@ source $1
 
 CHECKPOINT_PATH="/dataset/fd5061f6/english_data/finetune_checkpoints"
 
-N_GPU=2
 MASTER_PORT=$(shuf -n 1 -i 10000-65535)
 DISTRIBUTED_ARGS="--nproc_per_node ${N_GPU} --nnodes 1 --node_rank 0 --master_addr localhost --master_port $MASTER_PORT"
 
 DATESTR=$(date +"%m-%d-%H-%M")
 GRID_LOG=logs/grid_${EXPERIMENT_NAME}_${DATESTR}.txt
 
-bs=16
 for lr in 6e-6 1e-5 2e-5
 do
   for seed in 1234 5678 3456
   do
   HYPER=${lr}-${seed}
-  PER_GPU_BS=$((bs/N_GPU))
-  python -m torch.distributed.launch $DISTRIBUTED_ARGS finetune_gpt2.py \
-     --finetune \
-     --experiment-name ${EXPERIMENT_NAME}/${HYPER} \
-     --task ${TASK_NAME} \
-     --data-dir ${DATA_PATH} \
-     --save ${CHECKPOINT_PATH} \
-     --seq-length ${MAX_SEQ_LEN} \
-     --checkpoint-activations \
-     --batch-size 8 \
-     --eval-batch-size 16 \
-     --save-epoch 5 \
-     $MODEL_ARGS \
-     $TRAIN_ARGS \
-     $COMMON_ARGS \
-     --batch-size ${PER_GPU_BS} \
-     --epochs ${EPOCH_SINGLE} \
-     --lr-decay-style linear \
-     --lr ${lr} \
-     --seed ${seed} \
-     --overwrite \
-     2>&1 | tee logs/log-${EXPERIMENT_NAME}-${HYPER}.txt
+  PER_GPU_BS=$(($BATCH_SIZE/$N_GPU))
+  if [ ! -f runs/${EXPERIMENT_NAME}/${HYPER}/test_results.json ]; then
+    echo runs/${EXPERIMENT_NAME}/${HYPER}
+    python -m torch.distributed.launch $DISTRIBUTED_ARGS finetune_gpt2.py \
+       --finetune \
+       --experiment-name ${EXPERIMENT_NAME}/${HYPER} \
+       --task ${TASK_NAME} \
+       --data-dir ${DATA_PATH} \
+       --save ${CHECKPOINT_PATH} \
+       --seq-length ${MAX_SEQ_LEN} \
+       --checkpoint-activations \
+       --eval-batch-size 16 \
+       --save-epoch 5 \
+       $MODEL_ARGS \
+       $TRAIN_ARGS \
+       $COMMON_ARGS \
+       --batch-size ${PER_GPU_BS} \
+       --epochs ${EPOCH_SINGLE} \
+       --lr-decay-style linear \
+       --lr ${lr} \
+       --seed ${seed} \
+       --overwrite \
+       2>&1 | tee logs/log-${EXPERIMENT_NAME}-${HYPER}.txt
+  fi
   echo $lr $seed >> $GRID_LOG
   cat runs/${EXPERIMENT_NAME}/${HYPER}/results.json >> $GRID_LOG
   done
